+++ src/main/java/com/hsiaosiyuan/idea/ont/punica/OntPunica.java
@@ -5,8 +5,8 @@ import com.github.ontio.network.exception.ConnectorException;
 import com.github.ontio.network.exception.RpcException;
 import com.github.ontio.sdk.exception.SDKException;
 import com.hsiaosiyuan.idea.ont.run.OntRunCmdHandler;
-import com.hsiaosiyuan.idea.ont.ui.OntConsoleToolWindowFactory;
-import com.hsiaosiyuan.idea.ont.ui.OntNotifier;
+import com.hsiaosiyuan.idea.ont.run.OntConsoleToolWindowFactory;
+import com.hsiaosiyuan.idea.ont.run.OntNotifier;
 import com.hsiaosiyuan.idea.ont.util.OntSystemUtil;
 import com.intellij.execution.ExecutionException;
 import com.intellij.execution.configurations.GeneralCommandLine;
