+++ src/main/java/com/hsiaosiyuan/idea/ont/run/OntNotifier.java
@@ -1,4 +1,4 @@
-package com.hsiaosiyuan.idea.ont.ui;
+package com.hsiaosiyuan.idea.ont.run;
 
 import com.intellij.notification.*;
 import com.intellij.openapi.components.ServiceManager;
